requirements.txt
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
